chapter5/requirements.txt
fastapi
uvicorn
streamlit
streamlit-chat
python-dotenv
mcp
sentence-transformers>=2.7.0
chromadb
fastapi-mcp
google-genai
